requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
jinja2==3.1.6
python-multipart==0.0.20
python-dotenv==1.1.0
supabase==2.15.1